src/requirements.dev.txt
pytest==8.3.3
pytest-django==4.9.0
